+++ client_push_go17.go
@@ -1,0 +1,50 @@
+// +build go1.7
+
+package apns2
+
+import (
+	"bytes"
+	"context"
+	"encoding/json"
+	"fmt"
+	"io"
+	"net/http"
+)
+
+// Push sends a Notification to the APNs gateway. If the underlying http.Client
+// is not currently connected, this method will attempt to reconnect
+// transparently before sending the notification. It will return a Response
+// indicating whether the notification was accepted or rejected by the APNs
+// gateway, or an error if something goes wrong.
+//
+// Context with deadline allows to close long request when context timeout exceed. It can be nil, for back compatibility
+func (c *Client) PushWithCtx(n *Notification, ctx context.Context) (*Response, error) {
+	payload, err := json.Marshal(n)
+	if err != nil {
+		return nil, err
+	}
+
+	url := fmt.Sprintf("%v/3/device/%v", c.Host, n.DeviceToken)
+	req, _ := http.NewRequest("POST", url, bytes.NewBuffer(payload))
+
+	// add a context to request if exists
+	if ctx != nil {
+		req = req.WithContext(ctx)
+	}
+	setHeaders(req, n)
+	httpRes, err := c.HTTPClient.Do(req)
+	if err != nil {
+		return nil, err
+	}
+	defer httpRes.Body.Close()
+
+	response := &Response{}
+	response.StatusCode = httpRes.StatusCode
+	response.ApnsID = httpRes.Header.Get("apns-id")
+
+	decoder := json.NewDecoder(httpRes.Body)
+	if err := decoder.Decode(&response); err != nil && err != io.EOF {
+		return &Response{}, err
+	}
+	return response, nil
+}
